avoiding multi thread to lobby keepalive
